Moving profile view models in a separate folder (namespace).
