+++ FluentTerminal.App/Dialogs/QuickSshDialog.xaml
@@ -4,14 +4,14 @@
     xmlns:x="http://schemas.microsoft.com/winfx/2006/xaml"
     xmlns:d="http://schemas.microsoft.com/expression/blend/2008"
     xmlns:mc="http://schemas.openxmlformats.org/markup-compatibility/2006"
-    xmlns:viewModels="using:FluentTerminal.App.ViewModels"
     xmlns:controls="using:Microsoft.Toolkit.Uwp.UI.Controls"
     xmlns:enums="using:FluentTerminal.Models.Enums"
     xmlns:views="using:FluentTerminal.App.Views"
+    xmlns:profiles="using:FluentTerminal.App.ViewModels.Profiles"
     mc:Ignorable="d"
     x:Uid="QuickSshDialog"
     Title="Quick SSH"
-    d:DataContext="{d:DesignInstance viewModels:QuickSshViewModel, IsDesignTimeCreatable=False}"
+    d:DataContext="{d:DesignInstance profiles:QuickSshViewModel, IsDesignTimeCreatable=False}"
     PrimaryButtonText="OK"
     SecondaryButtonText="Cancel"
     PrimaryButtonClick="ContentDialog_PrimaryButtonClick"
